Run ucf for all scales

+++ ucf_sports/dim_14/features/optflow-feat-def.hpp
@@ -3,7 +3,7 @@ class opt_feat
 public:  
   inline opt_feat(const std::string in_path,
 		  const std::string in_actionNames,  
-		  const int in_scale_factor, 
+		  const float in_scale_factor, 
 		  const int in_shift,
 		  const int dim
   );
@@ -16,7 +16,7 @@ public:
   const std::string path;
   const std::string actionNames;
   
-  const int scale_factor;
+  const float scale_factor;
   const int shift;
   const int dim;
